Added error popup. Few bug fixes
When an error occurs, a popup box will be displayed with the error code.

Sound freq was decreased, this removed stutter caused when installing to the sd card (maxing the I/O).

Free storage space is updated after installing a game.

Nca placeholders are deleted if the install fails.

Added a few more font sizes.