Optimised ui, almost all textures are cached.
Textures (this includes buttons and text) are only updated if they need to be changed.

Previously with dynamic text, I would just recreate the texture every frame. However now in most cases, text is only updated every second, meaning that the texture is cached for 59 frames and then a new one is created.

This is *much* more efficient.

There's still quite a long way to go before it's perfect, however this a huge improvement from the previous version.