Add check for each storage / total size of selected partitions
It will calculate the total install size for nand AND sd card. This will be important soon once i add the *select install location* for per partition.

This also means that if the base game is 13gb and the update is 50mb. Say you have 100mb storage left, selecting to only install the update will now work.